+++ src/main/java/com/consulti/templatespringboot/models/PlanModel.java
@@ -13,12 +13,12 @@ public class PlanModel {
   @Id
   @GeneratedValue(strategy = GenerationType.AUTO)
   @Column(name = "plan_id", unique = true, updatable = false)
-  private int id;
+  private Long id;
 
   private String name;
 
   @Column(name = "user_id")
-  private int user_id;
+  private Long user_id;
 
   @Temporal(TemporalType.TIMESTAMP)
   @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
